+++ src/plugins/imageformats/jpeg/qjpeghandler.cpp
@@ -445,9 +445,9 @@ static bool write_jpeg_image(const QImage &sourceImage, QIODevice *device, int s
 
         jpeg_set_defaults(&cinfo);
 
-        float diffInch = qAbs(image.dotsPerMeterX()*2.54/100. - qRound(image.dotsPerMeterX()*2.54/100.))
+        qreal diffInch = qAbs(image.dotsPerMeterX()*2.54/100. - qRound(image.dotsPerMeterX()*2.54/100.))
                          + qAbs(image.dotsPerMeterY()*2.54/100. - qRound(image.dotsPerMeterY()*2.54/100.));
-        float diffCm = (qAbs(image.dotsPerMeterX()/100. - qRound(image.dotsPerMeterX()/100.))
+        qreal diffCm = (qAbs(image.dotsPerMeterX()/100. - qRound(image.dotsPerMeterX()/100.))
                         + qAbs(image.dotsPerMeterY()/100. - qRound(image.dotsPerMeterY()/100.)))*2.54;
         if (diffInch < diffCm) {
             cinfo.density_unit = 1; // dots/inch
